Fixed linter warning
