+++ integrationtests/resources/features/ansiblecluster/features_disabled_test.go
@@ -1,5 +1,6 @@
-//go:build (integration && featuretests) || allintegration
-// +build integration,featuretests allintegration
+//go:build !(allintegration || (integration && featuretests))
+// +build !allintegration
+// +build !integration !featuretests
 
 /*
  * Copyright 2018-2022, CS Systemes d'Information, http://csgroup.eu
@@ -17,15 +18,12 @@
  * limitations under the License.
  */
 
-package ansible_cluster
+package ansiblecluster
 
 import (
 	"testing"
-
-	"github.com/CS-SI/SafeScale/v22/integrationtests/helpers"
 )
 
-func Test_Ansible(t *testing.T) {
-	helpers.InSection("features").AddScenario(Ansible)
-	helpers.RunScenarios(t)
+func Test_Nop(t *testing.T) {
+	t.Skip("feature tests disabled")
 }
