+++ integrationtests/resources/features/ansiblecluster/playbook-failure-norequirements.zip
