+++ integrationtests/resources/features/ansiblecluster/scenarios.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package ansible_cluster
+package ansiblecluster
 
 import (
 	"fmt"
